Move SimpleCheckpointConfigurator into own module
