+++ great_expectations/checkpoint/__init__.py
@@ -11,7 +11,8 @@ from .actions import (
     UpdateDataDocsAction,
     ValidationAction,
 )
-from .checkpoint import Checkpoint, LegacyCheckpoint, SimpleCheckpointConfigurator
+from .checkpoint import Checkpoint, LegacyCheckpoint
+from .configurator import SimpleCheckpointConfigurator
 
 for module_name, package_name in [
     (".actions", "great_expectations.checkpoint"),
