+++ tests/checkpoint/test_simple_checkpoint.py
@@ -3,8 +3,8 @@ import pytest
 from great_expectations.checkpoint.checkpoint import (
     Checkpoint,
     CheckpointResult,
-    SimpleCheckpointConfigurator,
 )
+from great_expectations.checkpoint import SimpleCheckpointConfigurator
 from great_expectations.data_context.types.base import CheckpointConfig
